+++ cosmwasm/enclaves/ffi-types/src/lib.rs
@@ -9,7 +9,7 @@ pub use types::{
 };
 
 // 1 byte for length, 48 bytes for each potential encrypted seed
-pub const ENCRYPTED_SEED_SIZE: usize = 97;
+pub const ENCRYPTED_SEED_SIZE: u32 = 97;
 
 pub const SINGLE_ENCRYPTED_SEED_SIZE: usize = 48;
 pub const PUBLIC_KEY_SIZE: usize = 32;
